+++ packages/plainstudio/CHANGELOG.md
@@ -1,5 +1,40 @@
 # plainstudio
 
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  - plainstack@0.0.24
+
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - plainstack@0.0.23
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - plainstack@0.0.22
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - plainstack@0.0.21
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [0d4320d]
+  - plainstack@0.0.20
+
 ## 0.0.4
 
 ### Patch Changes